chore(functions): Added Unit Tests for the Basic Mapping Functions
Wanted to make sure that the lower level functions were working before moving on to higher level logic.

It appears all unit tests were passed, but a review of the logs indicates that we may need to be careful how errors are interpreted in higher level logic; it appears that "does not exist" might throw two errors?