feat(functions): Added Basic Mapping Functions (#4)
* fix(front-end): Updated Dependencies

* feat(functions): Added and Configured Firebase to Functions

* feat(functions): Added Basic Mapping Functions

These functions form the backbone of most future database operations for the video and contacts modules, so these were developed as modules first. This means that these trivial tasks are already done, and the future focus is just on the higher level logic.

Note, there are four main mappings that will be used:
- phone ---> userId: used to get user information from contacts
- email --> userId: used to get user information from contacts
- userId --> displayName: used to get display names for video conferencing
- userId --> photoUrl: used to get the profile photo for users for contacts, and for users without a web cam on in video conferencing

* chore(functions): Added Unit Tests for the Basic Mapping Functions

Wanted to make sure that the lower level functions were working before moving on to higher level logic.

It appears all unit tests were passed, but a review of the logs indicates that we may need to be careful how errors are interpreted in higher level logic; it appears that "does not exist" might throw two errors?